+++ src/components/atoms/waveform-select/square.tsx
@@ -2,15 +2,14 @@ export const Square = () => (
   <svg
     xmlns="http://www.w3.org/2000/svg"
     xmlSpace="preserve"
-    width="100%"
-    height="100%"
     viewBox="0 0 38 17"
+    className="size-auto h-auto w-[30px]"
   >
     <path
       d="M2.05 15.554v-14h10v14h10v-14h10v14h4"
       style={{
         fill: 'none',
-        stroke: '#000',
+        stroke: 'currentColor',
         strokeWidth: 1.5,
         strokeLinecap: 'round',
         strokeLinejoin: 'round',